move audit-related types to file event-types.d.ts instead of base-types.ts
