+++ packages/back-end/types/api.d.ts
@@ -1,4 +1,4 @@
-import { EventAuditUser } from "../src/events/base-types";
+import { EventAuditUser } from "../src/events/event-types";
 import { AuditInterface } from "./audit";
 import { ExperimentStatus } from "./experiment";
 import { OrganizationInterface } from "./organization";
